refactored logger
